🚚 [Core] move error types and formating functions to core/tools/error
